refactor: relocate navigation config
